+++ examples/react-hmr/rspack.config.mjs
@@ -54,11 +54,20 @@ export default (_, env) => {
         patterns: [{ from: 'manifest.json' }],
       }),
       new WebExtension({
-        background: { serviceWorkerEntry: 'background' },
+        background: { serviceWorkerEntry: 'background', serviceWorkerEntryOutput: 'sw.js' },
       }),
       isProduction ? null : new RefreshPlugin(),
     ].filter(Boolean),
-    optimization: { minimizer: [false] },
+    optimization: {
+      minimizer: [false],
+      // runtimeChunk: {
+      //   name: (entrypoint) => `runtime-${entrypoint.name}`,
+      // },
+      splitChunks: {
+        chunks: 'all',
+        // minSize: 1,
+      },
+    },
     experiments: { css: true },
   })
 }